@@ -128,10 +128,10 @@
     "lottie-react-native": "^5.1.3",
     "patch-package": "^6.2.0",
     "prop-types": "^15.7.2",
-    "react": "18.1.0",
+    "react": "18.2.0",
     "react-app-polyfill": "^2.0.0",
     "react-device-detect": "^2.0.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "18.2.0",
     "react-error-boundary": "^3.1.0",
     "react-helmet": "^6.1.0",
     "react-hook-form": "^7.35.0",
@@ -213,10 +213,10 @@
     "@types/jest-axe": "^3.5.5",
     "@types/lodash": "^4.14.182",
     "@types/node": "^14.14.41",
-    "@types/react": "^17.0.47",
-    "@types/react-dom": "^17.0.3",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9",
     "@types/react-helmet": "^6.1.2",
-    "@types/react-native": "^0.66.11",
+    "@types/react-native": "^0.70.8",
     "@types/react-native-calendars": "^1.1264.3",
     "@types/react-test-renderer": "^16.9.2",
     "@types/styled-components": "^5.1.19",
